fix(aot): use 64-bit CMP for i64 comparisons
The RA-path handler for .eq/.ne/.lt_*/.gt_*/.le_*/.ge_* selected 32-bit
vs 64-bit CMP based on inst.type, but inst.type is the destination
type (always i32 for comparisons), so i64 compares always used
cmpRegReg32 and ignored upper 32 bits.

Fix: frontend emits .type = .i64 for i64 compare/eqz ops. The handler
then correctly picks 64-bit CMP. The setcc+movzx result is still a
clean 0/1 byte so downstream consumers are unaffected.

Result: i64.json 36 -> 0 failures. AOT suite 14043 -> 14079
(83.4% -> 83.6%). Interp stays 20901/20901.

+++ src/compiler/frontend.zig
@@ -766,13 +766,13 @@ fn lowerFunction(func: *const types.WasmFunction, func_type: *const types.FuncTy
                     .i64_ge_u => .{ .ge_u = bin },
                     else => unreachable,
                 };
-                try ir_func.getBlock(current_block).append(.{ .op = ir_op, .dest = dest, .type = .i32 });
+                try ir_func.getBlock(current_block).append(.{ .op = ir_op, .dest = dest, .type = .i64 });
                 try vreg_stack.append(allocator, dest);
             },
             .i64_eqz => {
                 const val = safePop(&vreg_stack);
                 const dest = ir_func.newVReg();
-                try ir_func.getBlock(current_block).append(.{ .op = .{ .eqz = val }, .dest = dest, .type = .i32 });
+                try ir_func.getBlock(current_block).append(.{ .op = .{ .eqz = val }, .dest = dest, .type = .i64 });
                 try vreg_stack.append(allocator, dest);
             },
             .i64_clz, .i64_ctz, .i64_popcnt => {
